w3c-reports/12092014/mac/runned-tests/dom/nodes/Comment-constructor-expected.txt



PASS new Comment(): prototype chain 
PASS new Comment(): instanceof 
PASS new Comment(): no arguments 
FAIL new Comment(): undefined assert_equals: expected "" but got "undefined"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Comment-constructor.html:51:18
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/dom/nodes/Comment-constructor.html:49:7
forEach@[native code]
global code@http://localhost:8800/dom/nodes/Comment-constructor.html:47:18)
PASS new Comment(): null 
PASS new Comment(): 42 
PASS new Comment(): "" 
PASS new Comment(): "-" 
PASS new Comment(): "--" 
PASS new Comment(): "-->" 
PASS new Comment(): "<!--" 
PASS new Comment(): "\0" 
PASS new Comment(): "\0test" 
PASS new Comment(): "&amp;" 
PASS new Comment(): two arguments 
PASS new Comment() should get the correct ownerDocument across globals 
